+++ approver/Sources/AgentSecretApprover/Models/SessionBindingInfo.swift
@@ -5,6 +5,7 @@ public struct SessionBindingInfo: Codable, Equatable, Sendable {
         case mode
         case ancestorDepth = "ancestor_depth"
         case ancestorName = "ancestor_name"
+        case ancestorNames = "ancestor_names"
         case boundProcess = "bound_process"
         case creatorProcess = "creator_process"
     }
@@ -12,6 +13,7 @@ public struct SessionBindingInfo: Codable, Equatable, Sendable {
     public var mode: String
     public var ancestorDepth: Int?
     public var ancestorName: String?
+    public var ancestorNames: [String]
     public var boundProcess: SessionBindingProcess
     public var creatorProcess: SessionBindingProcess
 
@@ -20,12 +22,24 @@ public struct SessionBindingInfo: Codable, Equatable, Sendable {
         boundProcess: SessionBindingProcess,
         creatorProcess: SessionBindingProcess,
         ancestorDepth: Int? = nil,
-        ancestorName: String? = nil
+        ancestorName: String? = nil,
+        ancestorNames: [String] = []
     ) {
         self.mode = mode
         self.ancestorDepth = ancestorDepth
         self.ancestorName = ancestorName
+        self.ancestorNames = ancestorNames
         self.boundProcess = boundProcess
         self.creatorProcess = creatorProcess
+    }
+
+    public init(from decoder: Decoder) throws {
+        let container = try decoder.container(keyedBy: CodingKeys.self)
+        mode = try container.decode(String.self, forKey: .mode)
+        ancestorDepth = try container.decodeIfPresent(Int.self, forKey: .ancestorDepth)
+        ancestorName = try container.decodeIfPresent(String.self, forKey: .ancestorName)
+        ancestorNames = try container.decodeIfPresent([String].self, forKey: .ancestorNames) ?? []
+        boundProcess = try container.decode(SessionBindingProcess.self, forKey: .boundProcess)
+        creatorProcess = try container.decode(SessionBindingProcess.self, forKey: .creatorProcess)
     }
 }